requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
python-dotenv>=1.0.0
alpha_vantage>=2.3.1
requests>=2.31.0
duckdb>=0.9.0
streamlit>=1.32.0
plotly>=5.18.0 